Homogenize the deployment.
Use a standard deploy.bash. And fix master to use site.pp instead of it's unique init.pp.

Use the deploy.bash in the Dockerfile

Copy in content to the same place it would otherwise be if cloned. A little less efficient for docker caching, but much easier for manual installation.