Install reprepro from a prebuilt deb.
This brings and installs reprepro 5.1.1 from a prebuilt deb.
To build the Debian package use the dependencies/deb/build-reprepro.sh
script. I run it inside a container if it helps.

I'm not satisfied with the boundary between the reprepro component
module and the osrf::repo profile. That will need some work.