Use current Jenkins plugin API to generate puppet class for plugins.
This script can connect to a live jenkins instance and fetch plugin
data, then output a puppet class with the plugin versions and
dependencies specified.

This does not handle the job of configuring any plugins, that still
needs to be done manually.

There's lots of cleanup to be done with this script but I'm satisfied
enough for a Saturday afternoon.