Move jenkins configuration resources into dedicated module.
While each profile was independent it was perfectly fine to do a lot of heavy
lifting directly in the profile module, but now that we're trying to coinstall
all profiles we need to decompose a bit further.

This moves Jenkins master configuration from the profile::jenkins::master
module to the rosjenkins::manager module.  Along the way it also cleans up the
legacy jenkins_files module, moving the files and templates into the
rosjenkins::manager module.