Use puppetlabs docker fork.
Puppet has hard-forked from the previous Docker module we were using
which has fallen out of maintenance.

Since then, Docker have changed the location of their ubuntu repository
which means that we've been stuck with Docker 17.05.

Another fork of the previous docker module,
https://github.com/HackerHappyHour/hackerhappyhour-docker/ was also
under consideration but it appears to be targeting puppet 5.0+.

The selected fork only targets 4.0+ so it may also crash and burn for us
but the jump is smaller.