Update nvidia-docker2 package configuration.
* Use a local copy of the nvidia sources file.

Since http/https sources weren't introduced until Puppet 4.4, we can't
use them. And the wget approach is a decent fallback but I'd rather just
vendor the sources file and deal with linkrot of the apt repositories
rather than linkrot of the sources and the apt repositories.

* Notify Exec['apt_update'] from the apt package after the source has
  been added.

This should resolve race conditions between the apt source, apt key, and
apt update execution and means that the nvidia-docker2 package need only
depend on the proper source being configured.