Only install linux-aws when on an EC2 instance.
The linux-aws kernel is only necessary when using an EC2 instance.
In other systems that kernel may not be reasonable.

I'm not sure if other virtual systems will have specific kernel
requirements as I don't have any nvidia GPU + virtualizable systems to
test on. But that can be work for future PRs.