Prefer using require to before.
When two resources are both installed unconditionally. I arbitrarily
prefer declaring the dependency with `require` rather than `before`.

Using `before` is still needed in cases where a dependency needs to be
declared on a resource that is only installed conditionally. An example
is the linux-aws kernel needing to be available before nvidia-375 when
on an EC2 instance.