Use an updated puppet-jenkins module.
This bumps the forked jenkins module to include a patch I've just done
for a change in where the jenkins-cli jar is stored within recent
warfiles.

I haven't checked how far back this location is present for but with
issue #160 outstanding older versions of Jenkins won't ever get this
change. This is tested with only Jenkins 2.222.3